refactor(graph): extract D3 render logic into useGraphRender composable (Closes #35)
Direkt im Anschluss an die GraphPanel-Zerlegung (Issue #34) wandert die
renderGraph-Funktion samt D3-Lifecycle-State (currentSimulation,
linkLabelsRef, linkLabelBgRef), den drei Watchern (graphData deep,
showEdgeLabels, plus Erst-Render im Mount) und dem window.resize-
Listener aus GraphCanvas.vue in das neue Composable
frontend/src/composables/useGraphRender.js.

Composable-Vertrag:
- Eingaben: svgRef, containerRef, graphData (MaybeRefOrGetter),
  entityTypes (MaybeRefOrGetter), showEdgeLabels (Ref<boolean>)
- Ausgaben: selectedItem (Ref<object|null>), render (manual trigger)
- Lifecycle: onMounted hängt resize-Listener und ruft render im
  nextTick; onUnmounted räumt Listener und stoppt currentSimulation.
- Selection: D3-Click-Handler schreiben in selectedItem;
  GraphCanvas.closeDetailPanel kann den Ref von außen leeren.

D3-Verhalten unverändert: Drag-Threshold 3 px, Force-Parameter
(charge -400, collide 50, gravity 0.04), Zoom-ScaleExtent 0.1-4,
Tick-Updates, Edge-Labels-Visibility — alles 1:1.

Export-Funktionen (downloadGraphml, downloadSvg, downloadPng, printPdf)
bleiben in GraphCanvas.vue, weil sie DOM-nah am graphSvg-Ref hängen
und kein Render-Belang sind.

GraphCanvas.vue: 641 → 375 Zeilen.
useGraphRender.js: +317 Zeilen.

Akzeptanz Issue #35:
- useD3Graph.js oder ähnliches existiert ✓ (useGraphRender.js)
- Resize, Re-Render, Selection über klar definierte Schnittstellen ✓

Verifikation: npm run check 5/5 grün
- backend: 488 passed, 2 skipped
- frontend: 11 vitest passed, eslint clean, vite build (735 modules)

Closes #35
