+++ .gitignore
@@ -68,6 +68,7 @@ context.md
 !/docu/2026-05-01-issue34-sub21-graphhints-arbeitsprotokoll.md
 !/docu/2026-05-01-issue34-sub22-graphtoolbar-arbeitsprotokoll.md
 !/docu/2026-05-01-issue34-sub23-graphcanvas-arbeitsprotokoll.md
+!/docu/2026-05-01-issue35-usegraphrender-arbeitsprotokoll.md
 /docs/
 /Agora_design/
 /scripts/logs/
